+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add JMasterDetail component [#559](https://github.com/nbbrd/java-desktop-util/issues/559)
+
 ## [2.6.2] - 2026-04-09
 
 ### Fixed